pyproject.toml
[project]
name = "managebot"
version = "0.1.0"
description = ""
authors = [
    {name = "chuzeyu",email = "3343447088@qq.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "python-telegram-bot (>=22.2,<23.0)",
    "requests (>=2.32.4,<3.0.0)",
    "telethon (>=1.40.0,<2.0.0)",
    "python-socks[asyncio] (>=2.7.1,<3.0.0)",
]

[tool.poetry]
packages = [{include = "managebot", from = "src"}]

[tool.poetry.group.dev.dependencies]
python-dotenv = "^1.1.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"